backend/requirements.txt
# Jal Jeevan Mission - Backend Requirements

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
geoalchemy2==0.14.2
alembic==1.12.1

# MQTT
paho-mqtt==1.6.1
aiomqtt==1.2.1

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# ML/AI
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
joblib==1.3.2

# Real-time Processing
redis==5.0.1
celery==5.3.4

# Utilities
python-dateutil==2.8.2
pytz==2023.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
